+++ R/oauth-token.r
@@ -44,8 +44,10 @@ Token <- R6::R6Class("Token", list(
   credentials = NULL,
   params = NULL,
   cache_path = FALSE,
+  private_key = NULL,
 
   initialize = function(app, endpoint, params = list(), credentials = NULL,
+                        private_key = NULL,
                         cache_path = getOption("httr_oauth_cache")) {
     stopifnot(
       is.oauth_endpoint(endpoint) || !is.null(credentials),
@@ -57,6 +59,7 @@ Token <- R6::R6Class("Token", list(
     self$endpoint <- endpoint
     self$params <- params
     self$cache_path <- use_cache(cache_path)
+    self$private_key <- private_key
 
     if (!is.null(credentials)) {
       # Use credentials created elsewhere - usually for tests
@@ -137,11 +140,12 @@ Token <- R6::R6Class("Token", list(
 #' @export
 oauth1.0_token <- function(endpoint, app, permission = NULL,
                            as_header = TRUE,
+                           private_key = NULL,
                            cache = getOption("httr_oauth_cache")) {
   params <- list(permission = permission, as_header = as_header)
 
   Token1.0$new(app = app, endpoint = endpoint, params = params,
-    cache_path = cache)
+    private_key = private_key, cache_path = cache)
 }
 
 #' @export
@@ -149,7 +153,7 @@ oauth1.0_token <- function(endpoint, app, permission = NULL,
 Token1.0 <- R6::R6Class("Token1.0", inherit = Token, list(
   init_credentials = function(force = FALSE) {
     self$credentials <- init_oauth1.0(self$endpoint, self$app,
-      permission = self$params$permission)
+      permission = self$params$permission, private_key = self$private_key)
   },
   can_refresh = function() {
     FALSE
@@ -159,7 +163,7 @@ Token1.0 <- R6::R6Class("Token1.0", inherit = Token, list(
   },
   sign = function(method, url) {
     oauth <- oauth_signature(url, method, self$app, self$credentials$oauth_token,
-      self$credentials$oauth_token_secret)
+      self$credentials$oauth_token_secret, self$private_key)
     if (isTRUE(self$params$as_header)) {
       c(request(url = url), oauth_header(oauth))
     } else {